requirements.txt
langchain==0.3.25
langchain-community==0.3.23
langchain-groq==0.3.2
langchain-pinecone==0.2.0
pinecone-client==5.0.1
faiss-cpu==1.9.0
sentence-transformers==3.4.1
fastembed==0.4.2
streamlit==1.43.2
python-dotenv==1.1.0
pypdf==5.4.0
langchain-text-splitters==0.3.8
tiktoken==0.9.0
pytest==8.3.5